Removed unuseful widgets + new widget structure + management of working directory + xoppy_calc removed
