--- orangecontrib/xoppy/util/bin_run/README.txt
@@ -1,1 +1,0 @@
-This is the working directory for the xop executables
