+++ orangecontrib/xoppy/widgets/optics/icons/xraylib.png
