+++ orangecontrib/xoppy/widgets/source/xtube_w.py
@@ -1,25 +1,20 @@
-import sys
-import numpy as np
+import sys, os
+import numpy
 from PyQt4.QtGui import QIntValidator, QDoubleValidator, QApplication, QSizePolicy
 from PyMca5.PyMcaIO import specfilewrapper as specfile
 from orangewidget import gui
 from orangewidget.settings import Setting
-from oasys.widgets import widget
+from oasys.widgets.widget import OWWidget
+from oasys.widgets.exchange import DataExchangeObject
+from orangecontrib.xoppy.util.xoppy_util import locations
 
 try:
-    from orangecontrib.xoppy.util.xoppy_calc import xoppy_doc
+    from orangecontrib.xoppy.util.xoppy_util import xoppy_doc
 except ImportError:
     print("Error importing: xoppy_doc")
     raise
 
-try:
-    from orangecontrib.xoppy.util.xoppy_calc import xoppy_calc_xtube_w
-except ImportError:
-    print("compute pressed.")
-    print("Error importing: xoppy_calc_xtube_w")
-    raise
-
-class OWxtube_w(widget.OWWidget):
+class OWxtube_w(OWWidget):
     name = "xtube_w"
     id = "orange.widgets.dataxtube_w"
     description = "xoppy application to compute..."
@@ -30,7 +25,7 @@ class OWxtube_w(widget.OWWidget):
     category = ""
     keywords = ["xoppy", "xtube_w"]
     outputs = [{"name": "xoppy_data",
-                "type": np.ndarray,
+                "type": numpy.ndarray,
                 "doc": ""},
                {"name": "xoppy_specfile",
                 "type": str,
@@ -112,12 +107,12 @@ class OWxtube_w(widget.OWWidget):
             if sf.scanno() == 1:
                 #load spec file with one scan, # is comment
                 print("Loading file:  ",fileName)
-                out = np.loadtxt(fileName)
+                out = numpy.loadtxt(fileName)
                 print("data shape: ",out.shape)
                 #get labels
                 txt = open(fileName).readlines()
                 tmp = [ line.find("#L") for line in txt]
-                itmp = np.where(np.array(tmp) != (-1))
+                itmp = numpy.where(numpy.array(tmp) != (-1))
                 labels = txt[itmp[0]].replace("#L ","").split("  ")
                 print("data labels: ",labels)
                 self.send("xoppy_data",out)
@@ -133,6 +128,23 @@ class OWxtube_w(widget.OWWidget):
         print("help pressed.")
         xoppy_doc('xtube_w')
 
+
+def xoppy_calc_xtube_w(VOLTAGE=100.0,RIPPLE=0.0,AL_FILTER=0.0):
+    print("Inside xoppy_calc_xtube_w. ")
+
+    try:
+        with open("xoppy.inp","wt") as f:
+            f.write("%f\n%f\n%f\n"%(VOLTAGE,RIPPLE,AL_FILTER))
+    
+        command = os.path.join(locations.home_bin(), 'tasmip') + " < xoppy.inp"
+        print("Running command '%s' in directory: %s \n"%(command,locations.home_bin_run()))
+        print("\n--------------------------------------------------------\n")
+        os.system(command)
+        print("\n--------------------------------------------------------\n")
+
+        return "tasmip_tmp.dat"
+    except Exception as e:
+        raise e
