+++ orangecontrib/xoppy/widgets/tools/pickfile.py
