+++ orangecontrib/xoppy/widgets/tools/plot_simple.py
@@ -2,7 +2,7 @@ from PyQt4 import QtGui
 from PyQt4.QtGui import QDoubleValidator
 from orangewidget import widget, gui
 from orangewidget.settings import Setting
-import numpy as np
+import numpy
 import matplotlib.pyplot as plt
 from matplotlib.backends.backend_qt4agg import FigureCanvasQTAgg as FigureCanvas
 
@@ -20,7 +20,7 @@ class OWPlotSimple(widget.OWWidget):
     category = ""
     keywords = ["list", "of", "keywords"]
     inputs = [{"name": "xoppy_data",
-                "type": np.ndarray,
+                "type": numpy.ndarray,
                 "doc": "",
                 "handler": "do_plot" },
               {"name": "xoppy_specfile",
@@ -110,14 +110,14 @@ class OWPlotSimple(widget.OWWidget):
             pre_x2 = out[:, hcol[1]]
             pre_y = out[:, vcol]
 
-            x1 = np.unique(pre_x1)
-            x2 = np.unique(pre_x2)
+            x1 = numpy.unique(pre_x1)
+            x2 = numpy.unique(pre_x2)
             nx1 = len(x1)
             nx2 = len(x2)
             if (nx1*nx2 != len(pre_y)):
                 raise RuntimeError("Posible error in arrays: n1=%d, n2=%d, npoints=%d!=n1*n2."%(nx1,nx2,len(pre_y)))
 
-            y = np.zeros( (nx1,nx2) )
+            y = numpy.zeros( (nx1,nx2) )
             ij = 0
             for i in range(nx1):
                 for j in range(nx2):
@@ -127,7 +127,7 @@ class OWPlotSimple(widget.OWWidget):
                     ij += 1
             print("plot_simple: len x1,x2",len(x1),len(x2),len(x1)*len(x2))
             print("plot_simple: shape x1,x2,y",x1.shape,x2.shape,y.shape)
-            plt4 = plt.contour(x1,x2,y.T,levels=np.linspace(0,y.max(),100))
+            plt4 = plt.contour(x1,x2,y.T,levels=numpy.linspace(0,y.max(),100))
             cbar = plt.colorbar(plt4 , format="%.2f")
             cbar.ax.set_ylabel(labels[vcol])
             plt.ylabel(labels[hcol[1]])
@@ -142,7 +142,7 @@ class OWPlotSimple(widget.OWWidget):
 if __name__ == '__main__':
     app = QtGui.QApplication([])
     ow = OWPlotSimple()
-    # a = np.array([
+    # a = numpy.array([
     #     [  8.47091837e+04,   1.16210756e+12],
     #     [  8.57285714e+04,   1.10833975e+12],
     #     [  8.67479592e+04,   1.05700892e+12],
