--- orangecontrib/xoppy/widgets/viewers/__init__.py
@@ -1,18 +1,0 @@
-__author__ = 'srio'
-
-NAME = "Viewers"
-
-ID = "orangecontrib.xoppy.widgets.viewers"
-
-DESCRIPTION = """Widgets for x-ray oriented programs under python (xoppy)"""
-
-LONG_DESRIPTION = """
-This category contains widgets for calculating spectra of x-ray sources and reflectivities of optical elements (mirrors, multilayers, crystals, etc.).
-
-"""
-
-ICON = "icons/viewers.png"
-
-BACKGROUND = "#FFD39F"
-
-PRIORITY = 2004
