+++ orangecontrib/xoppy/widgets/xrayserver/__init__.py
@@ -9,12 +9,12 @@ Widgets for association rules.
 
 # Category description for the widget registry
 
-NAME = "GID_SL"
+NAME = "X-Ray Server"
 
-DESCRIPTION = "Widgets for X-Ray Server - GID_SL."
+DESCRIPTION = "Widgets for X-Ray Server "
 
 BACKGROUND = "#FFFFFF"
 
-ICON = "icons/gidsl.png"
+ICON = "icons/xrayserver.png"
 
-PRIORITY = 2007
+PRIORITY = 2004
