+++ orangecontrib/xoppy/widgets/xrayserver/icons/gidsl.png
