+++ orangecontrib/xoppy/widgets/xrayserver/icons/x0h.png
