--- orangecontrib/xoppy/widgets/xrayserver_x0h/__init__.py
@@ -1,20 +1,0 @@
-"""
-=========
-Associate
-=========
-
-Widgets for association rules.
-
-"""
-
-# Category description for the widget registry
-
-NAME = "X0h"
-
-DESCRIPTION = "Widgets for X-Ray Server - X0h"
-
-BACKGROUND = "#FFFFFF"
-
-ICON = "icons/x0h.png"
-
-PRIORITY = 2006
